Fix buffer resize bug (#16)
* Fix buffer resize bug

We were accidentally truncating the Vec<u8> values returned by
`Key#to_bytes()`. This patch should make sure that we are resizing
according to each individual key before joining them together into
a single Vec<u8>.

+++ codegen/rust/src/pipeline.rs
@@ -362,7 +362,7 @@ impl<'a> PipelineGenerator<'a> {
                 }
             }
 
-            let n = table.key.len() as usize;
+            let n = table.key.len();
             let table_type = quote! {
                 p4rs::table::Table::<
                     #n,
